repo::ebuild: use weak pointers for repo masters and trees
In order to avoid circular references between Arc pointers so freeing
repo objects via C bindings works as expected.

Note that calling the related masters() or trees() methods still returns
Arc pointers as we want the returned objects to be dereferenceable
without having to explicitly upgrade them. This is allowed because any
known repo must be registered in the config (which owns all the repo
objects) in order to be used.