initialize bash automatically for all test executables
Instead of having to manually determine if it's needed or not. This does
slow down running all tests a bit, but it's worth it overall to avoid
bash-related issues when forgetting to initialize it.

Note that all consumers using bash-related functionality will still have
to manually initialize bash. Currently this is done so consumers not
using bash or using external bash don't automatically run code that may
not work for them, e.g. running initialization for pkgcraft-bash causes
segfaults since it interacts poorly with the dynamic loading for
builtins.