pkgsh::builtins: rework builtin valid scope verification
Instead of only enabling builtins in their valid scopes, enable them for
all scopes and then verify the scope when they're called. This makes it
so bash never searches $PATH for a known builtin (as would have happened
before for un-enabled builtins in non-valid scopes) and instead should
always output the same types of error messages.