pkgcraft: revert to using rebuilt atom string for fmt::Display
Instead of storing a copy of the original. This means that an atom may
not have an string exactly the same as its input if it uses unordered
and/or collapsible use flags.

To support this in a better fashion if wanted in the future, a custom
OrderedSet type could be introduced that tracks the original elements
mapping to their new indexes in the underlying IndexSet and outputs them
(including duplicates and/or out of order) on iteration.