pkgcraft: use utf8 path handling for eapply
It's assumed since the majority of patch files are located either in the
tree or in the distfiles dir that they can be encoded in utf8.

Upstreams that bundle patches to be applied with non-utf8 compatible
file names are hopefully vanishingly rare and can be taught not to do
so.