scallop: add initial, basic forked process pool support
Note that this uses bash's SIGCHLD signal handler since it gets enabled
by default during command execution so trying to handle child processes
in scallop conflicts with waitpid() calls in bash's SIGCHLD handling.

In the future, this might install its own internal SIGCHLD handler and
then reset when the pool is dropped, but for now leveraging bash's
appears to work.