pkgcraft: globally check for functions that override PMS builtins
The overhead of this when running a full metadata generation on the
gentoo repo appears to be relatively negligible (<5% overhead) even when
errors exist.

To perform this sanely at a per-eclass-inherit level (and thus have
better errors when the erroneous functions are defined in eclasses)
requires parsing eclasses on repo init and caching the values.