pkgcraft: revert to unrestricted scope for assert, die, and nonfatal commands
It is possible that `die` is used in an eclass function called in global
ebuild scope so allow it everywhere as well as `assert` and `nonfatal`
since they go along with it.

Both `assert` and `nonfatal` are less useful in global scope since
external commands aren't allowed to be run, but builtins and pipelines
created using them can still be run so they are allowed anywhere `die`
can be used.