pkgcruft scan: use target restrictions builder
When creating target restrictions, path restrictions are now tried after
valid dep restrictions. To force a value that is both a valid dep and
path restriction to be treated as a path restriction, prefix the
relative path with "./".

For example, if the current working directory is the root directory for
a valid but unconfigured repo and the target "dev-python" is passed with
an existing "dev-python" category, the generated restriction will target
packages named "dev-python" and not packages within the "dev-python"
directory. If that behavior is wanted, force a path restriction by using
"./dev-python" instead.

This also re-adds support for specifying a target repo via --repo which
is mostly useful where dep restrictions are being used with a specific
target repo in mind (either its configured ID or an external repo path).
Without specifying a repo either the current directory will be used if
inside a valid repo or the default set of configured repos will be used.