pkgcruft: add new critical report level above error
Critical level reports are reserved for issues that invalidate or alter
other test report output such as sourcing failures or metadata errors.

This allows users to note that other reports may not be correct if a
critical level report exists for a given package, because that means the
package is skipped for other tests that rely on the ebuild being valid.