pkgcraft: add unordered ebuild package iterator support
Currently, iterating serially over an ebuild repo's packages constructs
one package at a time. Especially for packages lacking metadata, this
takes an excessive amount of time for newer, multi-core machines.

This adds an unordered iterator for ebuild repos that constructs
packages in parallel (including metadata generation) and returns them in
the order of completion.

Even when not generating any metadata this should be faster as it
parallelizes the regular construction of ebuild packages as well.