pk pkg manifest: add initial ebuild pkg manifesting support
This both downloads and hashes files in parallel. For those with a fast
network connection, using a higher concurrency setting via
-c/--concurrent is advised when targeting packages with a large number
of small distfiles.

This can be around 20-30x faster than `pkgdev manifest` when targeting
pkgs with a lot of small distfiles, e.g. rust-based ebuilds with many
crates, that must be downloaded and hashed.

Fixes #110.