pkgcruft: consistently skip pkg set tests if no pkgs exist
As the outdated todo note states, it's possible for a pkg set to be
empty for empty packages, however, currently we want iter_cpn_restrict()
to yield those values as well, e.g. it's useful in the case of
mismatched ebuild names that would otherwise be "empty" under standard
iteration.