pkgcraft: explicitly skip temp files instead of all hidden files for test repos
This tries to avoid race conditions where temporary metadata files are
created and deleted while copying repos due to running tests in
parallel.

It would probably be better to configure metadata generation to use an
external cache location and/or only copy test repos once per test suite
run, but this approach is easiest for now.