pkgcruft-git: add initial working pre-receive hook support
Currently it hackily relies on extracting the requested changes and
sending it as part of the scan request since it is difficult to
propagate commits from the temporary object directory used for incoming
changes during pre-receive hooks.

Furthermore, on success the HEAD commit is amended with the raw changes.
It would be possible to create entirely new commits, but if at all
possible it would be nice to send actual commit content without having
to extract it via spawning `git format-patch` or similar.