pkgcruft-git: rework push request to avoid unnecessary git checkouts
Git checkouts on large repos cause the most overhead compared to
scanning runs for simple commits. We try to avoid them at all costs by
not using a test branch approach.

In addition, if an error occurred or scanning failed the tree is
asynchronously reverted to its former state so the client doesn't wait
for it to occur.

This speeds up runtime almost 2x for simple commits.