scallop: fix passing through cargo llvm cov variables when testing
When the variables aren't passed through, the shell is initialized with
a blank environment causing the *.profraw files to be saved to the
current working directory.

Test that override the working directory cause the *.profraw files to be
saved outside the expected "target/llvm-cov-target" directory which then
aren't processed for the code coverage report.